requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
boto3==1.28.11
botocore==1.31.11
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dnspython==2.4.0
email-validator==2.0.0.post2
emoji==2.8.0
exceptiongroup==1.1.2
Flask==2.2.3
Flask-Cors==4.0.0
Flask-Dance==7.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
flask-mongoengine==1.0.0
Flask-WTF==1.1.1
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mongoengine==0.27.0
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.8
packaging==23.1
proto-plus==1.22.3
protobuf==4.24.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
stripe==6.0.0
tiktoken==0.4.0
tqdm==4.65.0
urllib3==1.26.16
URLObject==2.4.3
Werkzeug==2.3.6
WTForms==3.0.1
yarl==1.9.2
